# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Quest Room API
  version: 0.1.0
  description: |
    회원가입 없이 사용하는 공동 목표/반복 퀘스트 서비스의 MVP API 스펙.
    각 방은 최종 목표와 최종 목표 날짜를 가지며, 멤버들은 해당 날짜까지 함께 달린다.

    이 파일을 단일 진실의 원천(Single Source of Truth)으로 사용한다.
servers:
  - url: http://localhost:8080
    description: Local server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
...
    CreateRoomRequest:
      type: object
      required:
        - roomName
        - leaderNickname
        - finalGoal
        - finalGoalDate
        - timezone
      properties:
        roomName:
          type: string
          example: 100일 운동방
        leaderNickname:
          type: string
          maxLength: 24
          example: 시훈
        finalGoal:
          type: string
          example: 100일 동안 매일 운동 인증하기
        finalGoalDate:
          type: string
          format: date
          description: Final target date for the room.
          example: '2026-06-30'
        timezone:
          type: string
          description: IANA timezone identifier for the room.
          example: Asia/Seoul
...
    CreateRoomResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/CreateRoomData'
...
    InviteSummaryResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/InviteSummaryData'
    JoinInviteMemberRequest:
      type: object
      required:
        - nickname
      properties:
        nickname:
          type: string
          maxLength: 24
          example: 민지
...
    JoinInviteMemberResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/JoinInviteMemberData'
...
    RoomDashboardResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/RoomDashboardData'
    UpdateRoomRequest:
      type: object
      properties:
        roomName:
          type: string
          example: 100일 홈트방
        finalGoal:
          type: string
          example: 100일 동안 매일 홈트 인증하기
        finalGoalDate:
          type: string
          format: date
          example: '2026-07-31'
        dailyGoalCutoffPercent:
          type: integer
          minimum: 1
          maximum: 100
          description: |
            Daily goal cutoff percentage.
            The minimum required completion count is rounded from
            (active recurring quest count × dailyGoalCutoffPercent / 100).
          example: 60
      minProperties: 1
...
    RoomResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/RoomData'
...
    RoomListResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/RoomListData'
...
    RecurringQuestListResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/RecurringQuestListData'
...
paths:
...
  /v1/rooms:
    get:
      tags:
        - Rooms
      operationId: listRooms
      summary: 공개 방 목록 조회
      description: 메인 화면에서 보여줄 전체 방 목록을 최신 생성 순으로 조회한다.
      responses:
        '200':
          description: 방 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomListResponse'
    post:
      tags:
        - Rooms
      operationId: createRoom
      summary: 방 생성
      description: 방, 리더 멤버, 초기 세션을 생성한다. 방은 최종 목표 날짜까지 함께 달리는 단위를 의미한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: 방과 리더 멤버 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRoomResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/invites/{inviteToken}:
    get:
      tags:
        - Invites
      operationId: getInviteSummary
      summary: 초대 링크 정보 조회
      parameters:
        - name: inviteToken
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 초대 링크 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteSummaryResponse'
        '404':
          description: 초대 링크를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/invites/{inviteToken}/members:
    post:
      tags:
        - Invites
        - Members
      operationId: joinInviteAsMember
      summary: 초대 링크로 최초 참가
      parameters:
        - name: inviteToken
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinInviteMemberRequest'
      responses:
        '201':
          description: 멤버 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinInviteMemberResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 초대 링크를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 닉네임 중복 등 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/{roomId}:
    get:
      tags:
        - Rooms
      operationId: getRoomDashboard
      summary: 방 대시보드 조회
      description: 현재 인증된 멤버 기준으로 방의 공동 목표와 오늘 수행 상태를 조회한다.
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 방 대시보드 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDashboardResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Rooms
      operationId: updateRoom
      summary: 방 이름, 최종 목표 또는 최종 목표 날짜 수정
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: 방 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 리더 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/{roomId}/members/{memberId}:
    delete:
      tags:
        - Members
      operationId: kickRoomMember
      summary: 방 멤버 강퇴
      description: |
        리더가 방의 일반 멤버를 강퇴한다.
        리더 자신이나 다른 리더는 강퇴할 수 없다.
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 멤버 강퇴 성공
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 리더 권한이 없거나 강퇴할 수 없는 대상
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방 또는 멤버를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/{roomId}/recurring-quests:
    get:
      tags:
        - RecurringQuests
      operationId: listRecurringQuests
      summary: 반복 퀘스트 목록 조회
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 반복 퀘스트 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringQuestListResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - RecurringQuests
      operationId: createRecurringQuest
      summary: 반복 퀘스트 생성
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecurringQuestRequest'
      responses:
        '201':
          description: 반복 퀘스트 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringQuestResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 리더 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/{roomId}/recurring-quests/{questId}:
    patch:
      tags:
        - RecurringQuests
      operationId: updateRecurringQuest
      summary: 반복 퀘스트 수정
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: questId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecurringQuestRequest'
      responses:
        '200':
          description: 반복 퀘스트 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringQuestResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 리더 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방 또는 퀘스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/{roomId}/completions:
    get:
      tags:
        - Completions
      operationId: listCompletions
      summary: 날짜 범위별 완료 기록 조회
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: questId
          in: query
          required: false
          schema:
            type: string
        - name: memberId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 완료 기록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletionListResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/{roomId}/daily-status/{date}:
    get:
      tags:
        - Rooms
        - Completions
      operationId: getRoomDailyStatus
      summary: 특정 날짜의 방 상태 조회
      description: 방의 특정 날짜 기준 반복 퀘스트 목록과 완료 기록을 함께 조회한다.
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 날짜별 방 상태 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDailyStatusResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/{roomId}/recurring-quests/{questId}/completions/{date}:
    put:
      tags:
        - Completions
      operationId: putMyQuestCompletion
      summary: 내 반복 퀘스트 완료 처리
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: questId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutCompletionRequest'
      responses:
        '200':
          description: 완료 처리 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletionResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방 또는 퀘스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Completions
      operationId: deleteMyQuestCompletion
      summary: 내 완료 기록 취소
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: questId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '204':
          description: 완료 기록 취소 성공
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방, 퀘스트 또는 완료 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/rooms/{roomId}/invite-token:rotate:
    post:
      tags:
        - Rooms
        - Invites
      operationId: rotateInviteToken
      summary: 초대 링크 재발급
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 초대 토큰 재발급 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateInviteTokenResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 리더 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 방을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'